Check for too large message prefix
instead of message header, thus making the branch code consitent
with old reply_header_max_size checks. Removed HttpHeader::tooLarge():
we do not need such general method now because it is simpler to
detect large headers with the recently added HttpReply::prefixLen().

Also fixed unit tests.

+++ src/HttpReply.h
@@ -133,7 +133,7 @@ public:
     size_t parseTerminatedPrefix(const char *, size_t);
 
     /// \returns the length of status line + headers + crlf
-    int prefixLen() const;
+    size_t prefixLen() const;
 
 private:
     /** initialize */
